IMPORTANT: Managed to call RS AMG for pressure!
OK, the call was done only from GPU version, if compiled by GNU Fortran.

A number of interventions were needed to achieve this:

1. All declarations of double precision in Amg1r5 source have been changed
   to real, letting the compiler turn the reals to double precision by a
   compiler option.  Leaving double precision in declarations, lead to
   extended (real(16)) floating point numbers.
2. Following the above, all numbers which were written in double precision
   mode (such as 0.d0 0.1d0 and alike, have been changed to real versions
   (such as 0.0 and 0.1).
3. Also, all cast operators were changed from dble to real.
4. New keyword has been introduced, "rs_amg" to allow a user to pick the
   Ruge-Stueben AMG solver.
5. In the GPU version, Var_Mod was extended by the field "solver", to hold
   this piece of information.
6. Quite important: matrices are now stored in format suggested by Ruge and
   Stueben, diagonal is stored firstvi control!  That lead to changes in:
   Sources/Process/Gpu/Sparse_Mod/Create_Sparse.f90
   Sources/Process/Cpu/Matrix_Mod/Create_Matrix.f90;

   but also in the following:

   Sources/Process/Cpu/Native_Mod/Prec_Form.f90
   Sources/Process/Cpu/Native_Mod/Prec_Solve.f90

   since matrix-vector product for upper and lower portion of the matrices
   had to be adjusted.
7. A couple of internal changes were also done inside the Amg_Mod to allow
   better interaction with the outside world.

On branch multiple_scalars_gpu_bojan

modified:   Documentation/Syntax/.vim/syntax/control.vim
modified:   Documentation/all_control_keywords
modified:   Sources/Process/Cpu/Matrix_Mod/Create_Matrix.f90
modified:   Sources/Process/Cpu/Native_Mod/Prec_Form.f90
modified:   Sources/Process/Cpu/Native_Mod/Prec_Solve.f90
modified:   Sources/Process/Gpu/Field_Mod.f90
modified:   Sources/Process/Gpu/Process_Mod/Compute_Pressure.f90
modified:   Sources/Process/Gpu/Process_Mod/Compute_Pressure.fpp
modified:   Sources/Process/Gpu/Read_Controls_Mod/Native_Solvers.f90
modified:   Sources/Process/Gpu/Sparse_Mod/Create_Sparse.f90
modified:   Sources/Process/Gpu/Var_Mod.f90
modified:   Sources/Shared/Amg_Mod.f90
modified:   Sources/Shared/Amg_Mod/0_Drivers/Amg1r5.f90
modified:   Sources/Shared/Amg_Mod/1_Setup.f90
modified:   Sources/Shared/Amg_Mod/1_Setup/1_Check_Matrix_Properties.f90
modified:   Sources/Shared/Amg_Mod/1_Setup/2_Coarsening.f90
modified:   Sources/Shared/Amg_Mod/1_Setup/2_Coarsening/1_Row_Sort.f90
modified:   Sources/Shared/Amg_Mod/1_Setup/2_Coarsening/3_Interpolation_Weights.f90
modified:   Sources/Shared/Amg_Mod/1_Setup/2_Coarsening/4_Define_Operators.f90
modified:   Sources/Shared/Amg_Mod/1_Setup/2_Coarsening/5_Truncate_Operator.f90
modified:   Sources/Shared/Amg_Mod/2_First_Guess.f90
modified:   Sources/Shared/Amg_Mod/3_Solve.f90
modified:   Sources/Shared/Amg_Mod/3_Solve/1_Calculate_Residual.f90
modified:   Sources/Shared/Amg_Mod/3_Solve/2_Backup_U.f90
modified:   Sources/Shared/Amg_Mod/3_Solve/3_One_Cycle.f90
modified:   Sources/Shared/Amg_Mod/3_Solve/3_One_Cycle/1_Solve_On_Coarsest_Level.f90
modified:   Sources/Shared/Amg_Mod/3_Solve/3_One_Cycle/1_Solve_On_Coarsest_Level/1_Cg_On_Coarsest_Level.f90
modified:   Sources/Shared/Amg_Mod/3_Solve/3_One_Cycle/1_Solve_On_Coarsest_Level/2_Bicg_On_Coarsest_Level.f90
modified:   Sources/Shared/Amg_Mod/3_Solve/3_One_Cycle/2_Normalize_U.f90
modified:   Sources/Shared/Amg_Mod/3_Solve/3_One_Cycle/3_Gauss_Seidel_Sweep.f90
modified:   Sources/Shared/Amg_Mod/3_Solve/3_One_Cycle/5_Set_U_To_Zero.f90
modified:   Sources/Shared/Amg_Mod/3_Solve/3_One_Cycle/6_Restrict_Residuals.f90
modified:   Sources/Shared/Amg_Mod/3_Solve/3_One_Cycle/7_Scale_Solution.f90
modified:   Sources/Shared/Amg_Mod/3_Solve/3_One_Cycle/8_Interpolate_Correction.f90
modified:   Sources/Shared/Amg_Mod/3_Solve/4_Cg_Step.f90
modified:   Sources/Shared/Amg_Mod/3_Solve/4_Cg_Step/1_Cg_Alpha.f90
modified:   Sources/Shared/Amg_Mod/3_Solve/4_Cg_Step/2_Cg_Epsilon.f90
modified:   Sources/Shared/Amg_Mod/4_Wrkcnt.f90
modified:   Sources/Shared/Amg_Mod/8_Utilities/Get_Integer_Digits.f90
modified:   Sources/Shared/Amg_Mod/8_Utilities/Random_0_To_0p1.f90
modified:   Sources/Shared/Control_Mod/Native/Solver_For_Pressure.f90

+++ Sources/Shared/Amg_Mod/1_Setup.f90
@@ -9,13 +9,13 @@
 !------------------------------------------------------------------------------!
   implicit none
 !---------------------------------[parameters]---------------------------------!
-  class(Amg_Type)  :: Amg
-  integer          :: nnu, levelx
-  double precision :: a(:), u(:)
-  integer          :: ia(:), ja(:)
-  integer          :: iw(:), icg(:), ifg(:)
-  integer          :: levels
-  integer          :: iwork(:), jtr(:)
+  class(Amg_Type) :: Amg
+  integer         :: nnu, levelx
+  real            :: a(:), u(:)
+  integer         :: ia(:), ja(:)
+  integer         :: iw(:), icg(:), ifg(:)
+  integer         :: levels
+  integer         :: iwork(:), jtr(:)
 !-----------------------------------[locals]-----------------------------------!
   integer :: i, n_digits
   integer :: digit(AMG_MAX_LEVELS)
