+++ Sources/Shared/Amg_Mod/3_Solve/3_One_Cycle/2_Normalize_U.f90
@@ -5,12 +5,12 @@
 !------------------------------------------------------------------------------!
   implicit none
 !---------------------------------[parameters]---------------------------------!
-  class(Amg_Type)  :: Amg
-  integer          :: level
-  double precision :: u(:)
+  class(Amg_Type) :: Amg
+  integer         :: level
+  real            :: u(:)
 !-----------------------------------[locals]-----------------------------------!
-  double precision :: fac
-  integer          :: i
+  real    :: fac
+  integer :: i
 !------------------------------------[save]------------------------------------!
   save  ! this is included only as a precaution as Ruge-Stueben had it
 !==============================================================================!
